test/requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cloudscraper==1.2.71
h11==0.14.0
idna==3.4
-e git+ssh://git@github.com/Bizzaro/nbdb-python-api.git@146f486db45a8d506104b55b378158d7f00c1f76#egg=nbdapi
outcome==1.2.0
pycparser==2.21
pyparsing==3.1.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
undetected-chromedriver==3.5.3
urllib3==2.0.6
websockets==11.0.3
wsproto==1.2.0
